DEV: Fix flaky spec
The issue was that when we were moving the post using posts[1,2]
the ordering was not always consistent, so sometimes we moved post
number 1 for the topic and sometimes post number 2.

If post number 1 was moved, the PostMover just creates a copy of it
in the destination topic, as the OP of the topic cannot be moved.

If post number 2 was moved, then PostMover _actually_ moves it to
the destination topic, and thus leaves only one post behind in
the original topic, which causes the original topic to be closed
and makes the spec fail.

This change ensures that post number 1 is the one we are moving
in the test, making sure the order is constant.