DEV: Rename `has_one :category_setting` as it clashes with core
What is the problem?

Discourse core already declares a `has_one :category_setting`
association so this plugin is overriding that association which is not
what we want.

What is the solution?

Since this plugin adds a category setting model of its own, we can
simply rename the association to avoid the conflict with core.